+++ frontend/app/questions/style-cell.tsx
@@ -1,9 +1,8 @@
 import { Chip } from '@nextui-org/chip';
-import { Tooltip } from '@nextui-org/tooltip';
-import { DeleteIcon } from './assets/DeleteIcon';
-import QuestionDescriptionModal from './question-decription-modal';
+import QuestionDescriptionModal from './descriptionModal';
 import { Category, ComplexityToColor, Question } from '../types/question';
 import { Key } from 'react';
+import DeleteConfirmationModal from './deleteConfirmationModal';
 
 interface StyleCellProps {
   item: Question & { id: number };
@@ -32,11 +31,7 @@ const StyleCell: React.FC<StyleCellProps> = ({ item, columnKey }) => {
       return (
         <div className="relative flex items-center gap-5">
           <QuestionDescriptionModal title={item.title} description={item.description} />
-          <Tooltip color="danger" content="Delete question">
-            <span className="text-lg text-danger cursor-pointer active:opacity-50">
-              <DeleteIcon />
-            </span>
-          </Tooltip>
+          <DeleteConfirmationModal title={item.title} />
         </div>
       );
